Fix personal access to data #19

+++ server/src/routes/api/baby/deleteFeedingSchedule.js
@@ -9,6 +9,38 @@ module.exports = async (req, res) => {
   const { mealId } = req.params;
 
   try {
+    // Decode the token to get the user ID
+    const authHeader = req.headers.authorization;
+    if (!authHeader) {
+      return res.status(401).json({
+        status: "error",
+        error: {
+          message: "No authorization token provided",
+        },
+      });
+    }
+
+    const user_id = await getUserId(authHeader);
+    if (!user_id) {
+      return res.status(404).json({
+        status: "error",
+        error: {
+          message: "User not found",
+        },
+      });
+    }
+
+   // {CHECK OWNERSHIP of BABY}
+    // Verify user has access to this baby
+    const hasAccess = await checkBabyBelongsToUser(baby_id, user_id);
+    if (!hasAccess) {
+      return res
+        .status(403)
+        .json(
+          createErrorResponse("Not authorized to access this baby profile")
+        );
+    }
+    
     const deletedFeedingSchedules = await pool.query(
       "DELETE FROM feedingschedule WHERE feeding_schedule_id = $1 RETURNING *",
       [mealId]
